[REFACTOR] Day enum 변경에 따른 리팩토링
